requirements.txt
antlr4-python3-runtime==4.9.3
hydra-core==1.3.2
importlib_resources==6.4.0
nptyping==2.5.0
numpy==1.24.4
omegaconf==2.3.0
packaging==24.0
PyYAML==6.0.1
typing_extensions==4.11.0
zipp==3.18.2
